fix: uint16_t overflow in currentDegrees defeats OVH protection
On STM8/SDCC, int is 16-bit. The expression
  (MAX_HEAT - MIN_HEAT) * (adcVal - adcMinRT)
is evaluated as unsigned int * unsigned int (both 16-bit), which
overflows uint16_t when adcVal - adcMinRT > 163.

When the overflow wraps the product to a small number, currentDegrees
is computed as 0-5 degrees even though the sensor reads an out-of-range
value (e.g. tip removed, sensor noise, or adcMaxRT calibrated to 500).
The OVH hard-limit check (currentDegrees > MAX_SAFE_TEMP = 480) then
never fires, leaving the heater running at full power.

Fix: cast to uint32_t before the multiplication so the intermediate
value is computed in 32-bit arithmetic. With uint32_t, a worst-case
adcVal of 1023 and adcMinRT=1 gives 400 * 1022 = 408800, which fits
easily and produces a currentDegrees > 480, correctly triggering OVH.

+++ main.c
@@ -159,7 +159,10 @@ void mainLoop(void)
 
     // Degrees value — use per-tip calibrated ADC range
     adcVal = (adcVal < _eepromData.adcMinRT) ? _eepromData.adcMinRT : adcVal;
-    int16_t currentDegrees = (MAX_HEAT - MIN_HEAT) * (adcVal - _eepromData.adcMinRT) / (_eepromData.adcMaxRT - _eepromData.adcMinRT);
+    // Cast to uint32_t before the multiply: on STM8/SDCC int is 16-bit, so
+    // (MAX_HEAT-MIN_HEAT)*N overflows uint16_t when N>163, silently wrapping
+    // currentDegrees to a small value and defeating the OVH protection check.
+    int16_t currentDegrees = (int16_t)((uint32_t)(MAX_HEAT - MIN_HEAT) * (adcVal - _eepromData.adcMinRT) / (_eepromData.adcMaxRT - _eepromData.adcMinRT));
     currentDegrees += _eepromData.calibrationValue;
 
     // ER1: short on sensor  (ADC well below the calibrated cold point)
